Set triggers for Actions and added keys to keyList
Initilized Action triggers in Logic and InputKeys in Input so in theory Logic can get the state of an Action's key without knowing what key it's checking.
Currently not working though cause the tirggers aren't being set in the contructor for some reason and I'm not sure why yet. Looks like the contructor is being called in HeadHancho but it's not doing any of the settings stuff