Partially implemented the update functions for Objects
Update functions for Objects are now called when the Object is created and deleted. In this case it's not actually when the Object is created/deleted in memory (as when the Object is created it may not have had the chance to be assigned an update function), but they are called when they are added or removed from the ObjectList in RM.

Update during runtime is not implemented yet.