Added Space Invader level and some other Object cleanup
Renamed some assets to be more obvious
Added some assets for space invader level
Added check fo object collision to check if objects are null. This is used to fix issue where a collision causes other objects to be deleted in the middle of an all object collision check (like moving to a new level). Might change to include zombie bool for levels instead but not sure.
Added additional Object types for space invader Objects
Named enum in ObjectType to allow it to be used as a function parameter
Cleanup on player movement (don't need it saying direction since it works)
Changed Object creation in levels to get spriteID instead of hardcoded numbers (needed for when assets are added./removed)
Added level update (creation only), player creation, and bullet creation functions for space invader level