Split moveAllEnemies() into moveEnemiesCheck and a new moveAllEnemies
Enemies now move across the screen properly with the two functions.
moveEnemiesCheck checks which direction the enemies should all move in (based on the enemiesMoveRight bool and if enemies would go offscreen) then it calls moveAllEnemies which just moves the enemies with given direction.
Also added getEdgeEnemy which gets the enemy on the outermost edge (or one of them if there are multiple), which is used to check if the enemies are going to go offscreen